+++ src/object/types.ts
@@ -29,9 +29,9 @@ export type DotNotationKeyStrict<T> =
 	T extends AdvancedTypes ? never
 	: T extends StrictObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends StrictObject ?
-				`${K}` | `${K}.${DotNotationKey<NonNullable<T[K]>>}`
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends StrictObject ?
+				`${K}` | `${K}.${DotNotationKey<T[K]>}`
 			:	`${K}`;
 		}[keyof T & string]
 	:	never;
@@ -41,9 +41,9 @@ export type DotNotationKey<T> =
 	T extends AdvancedTypes ? never
 	: T extends GenericObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends GenericObject ?
-				`${K}` | `${K}.${DotNotationKey<NonNullable<T[K]>>}`
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends GenericObject ?
+				`${K}` | `${K}.${DotNotationKey<T[K]>}`
 			:	`${K}`;
 		}[keyof T & string]
 	:	never;
@@ -52,8 +52,8 @@ export type DotNotationKey<T> =
 export type KeyForArray<T> =
 	T extends GenericObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends Array<unknown> ? K
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends Array<unknown> ? K
 			: never;
 		}[keyof T & string]
 	:	never;
@@ -63,9 +63,9 @@ export type KeyForObject<T> =
 	T extends AdvancedTypes ? never
 	: T extends GenericObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends GenericObject ?
-				NonNullable<T[K]> extends AdvancedTypes ?
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends GenericObject ?
+				T[K] extends AdvancedTypes ?
 					never
 				:	K
 			:	never;
@@ -77,11 +77,10 @@ export type NestedKeyString<T> =
 	T extends AdvancedTypes ? never
 	: T extends GenericObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends string ? K
-			: NonNullable<T[K]> extends GenericObject ?
-				`${K}.${NestedKeyString<NonNullable<T[K]>>}`
-			:	never;
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends string ? K
+			: T[K] extends GenericObject ? `${K}.${NestedKeyString<T[K]>}`
+			: never;
 		}[keyof T & string]
 	:	never;
 
@@ -90,11 +89,10 @@ export type NestedPrimitiveKey<T> =
 	T extends AdvancedTypes ? never
 	: T extends GenericObject ?
 		{
-			[K in keyof T & string]: NonNullable<T[K]> extends Function ? never
-			: NonNullable<T[K]> extends NormalPrimitive ? K
-			: NonNullable<T[K]> extends GenericObject ?
-				`${K}.${NestedPrimitiveKey<NonNullable<T[K]>>}`
-			:	never;
+			[K in keyof T & string]: T[K] extends Function ? never
+			: T[K] extends NormalPrimitive ? K
+			: T[K] extends GenericObject ? `${K}.${NestedPrimitiveKey<T[K]>}`
+			: never;
 		}[keyof T & string]
 	:	never;